refactor(core): add `packages/core/sanitization` bazel target
1. Breaking up monolithic build targets into smaller builds is desirable.
2. This is motivated by code duplication between `packages/core` and `packages/compiler`. By breaking `packages/core` into smaller sub-builds it will be possible for the `packages/compiler` to directly depend on `packages/core/src/sanitization` which will than allow removal of code duplication.

While creating the smaller bundle it was discovered that these two files needed to be moved:
- `packages/core/src/sanitization/sanitizer.ts`: This file was depending on DI and as such it would prevent `packages/core/src/sanitization` from being a leaf target. Moved to `packages/cor/src/sanitizer/sanitizer.ts`
- `packages/core/src/sanitization/sanitization.ts`: This file was depending on render3 directory and as such it would prevent `packages/core/src/sanitization` from being a leaf target. Moved to `packages/cor/src/render3/sanitization.ts`. The contents of this file contains Ivy instructions and so the new location is a better fit for it.
