+++ packages/core/src/core.ts
@@ -35,7 +35,7 @@ export {ErrorHandler} from './error_handler';
 export * from './core_private_export';
 export * from './core_render3_private_export';
 export {SecurityContext} from './sanitization/security';
-export {Sanitizer} from './sanitization/sanitizer';
+export {Sanitizer} from './sanitizer/sanitizer';
 export * from './codegen_private_exports';
 
 import {global} from './util/global';
