+++ packages/core/src/core_render3_private_export.ts
@@ -205,7 +205,7 @@ export {
   ɵɵsanitizeUrl,
   ɵɵsanitizeResourceUrl,
   ɵɵsanitizeUrlOrResourceUrl,
-} from './sanitization/sanitization';
+} from './render3/sanitization';
 
 export {
   bypassSanitizationTrustHtml as ɵbypassSanitizationTrustHtml,
