+++ packages/core/src/render3/component.ts
@@ -10,7 +10,7 @@
 // correctly implementing its interfaces for backwards compatibility.
 import {Type} from '../core';
 import {Injector} from '../di/injector';
-import {Sanitizer} from '../sanitization/sanitizer';
+import {Sanitizer} from '../sanitizer/sanitizer';
 import {assertDataInRange} from '../util/assert';
 
 import {assertComponentType} from './assert';
