+++ packages/core/src/sanitizer/readme.md
