+++ packages/core/src/view/types.ts
@@ -15,8 +15,8 @@ import {QueryList} from '../linker/query_list';
 import {TemplateRef} from '../linker/template_ref';
 import {ViewContainerRef} from '../linker/view_container_ref';
 import {Renderer2, RendererFactory2, RendererType2} from '../render/api';
-import {Sanitizer} from '../sanitization/sanitizer';
 import {SecurityContext} from '../sanitization/security';
+import {Sanitizer} from '../sanitizer/sanitizer';
