+++ packages/core/test/render3/integration_spec.ts
@@ -14,9 +14,9 @@ import {MONKEY_PATCH_KEY_NAME} from '../../src/render3/interfaces/context';
 import {RenderFlags} from '../../src/render3/interfaces/definition';
 import {RElement, Renderer3, RendererFactory3, domRendererFactory3} from '../../src/render3/interfaces/renderer';
 import {CONTEXT, HEADER_OFFSET} from '../../src/render3/interfaces/view';
-import {ɵɵsanitizeUrl} from '../../src/sanitization/sanitization';
-import {Sanitizer} from '../../src/sanitization/sanitizer';
+import {ɵɵsanitizeUrl} from '../../src/render3/sanitization';
 import {SecurityContext} from '../../src/sanitization/security';
+import {Sanitizer} from '../../src/sanitizer/sanitizer';
 
 import {NgIf} from './common_with_def';
 import {ComponentFixture, MockRendererFactory, renderToHtml} from './render_util';
